build mingw-w64
- build and test pass mingw32, x86_64-w64-mingw32(tdm64)
- newlisp.h: INT/UINT use stdint.h
- nl-import.c: suppress callback cast
- nl-math.c: correct compare cast
- nl-sock.c: avoid UINT conflict
- nl-string.c: fix MS-I64 format type
- win32-path.c: temporary fix struct _stat
- win32-util.c: fix cast between newlisp_UINT and HANDLE

fix printf value format

+++ nl-math.c
@@ -569,7 +569,7 @@ double floatNum = 0.0;
 if(param->type == CELL_FLOAT)
     return(*(double *)&param->aux);
 else if(param->type == CELL_LONG)
-    floatNum = (long)param->contents;
+    floatNum = (long)param->contents; /* INT ? */
 else if(param->type == CELL_INT64)
     floatNum = *(INT64 *)&param->aux;
 #ifdef BIGINT
@@ -1066,8 +1066,8 @@ switch(left->type)
 #endif 
     case CELL_LONG:
     default:
-        if((long)left->contents > (long)right->contents) return(1);
-        if((long)left->contents < (long)right->contents) return(-1);
+        if((INT)left->contents > (INT)right->contents) return(1);
+        if((INT)left->contents < (INT)right->contents) return(-1);
         break;  
     }
